[feat] MyPage ViewModel 구현
